+++ client/jetbrains/src/main/java/com/sourcegraph/config/CodyApplicationService.java
@@ -89,10 +89,7 @@ public class CodyApplicationService
     // configuring enterpriseAccessToken overrides the deprecated accessToken field
     String configuredEnterpriseAccessToken =
         StringUtils.isEmpty(enterpriseAccessToken) ? accessToken : enterpriseAccessToken;
-    // defaulting to SRC_ACCESS_TOKEN env if nothing else was configured
-    return configuredEnterpriseAccessToken == null
-        ? System.getenv("SRC_ACCESS_TOKEN")
-        : configuredEnterpriseAccessToken;
+    return configuredEnterpriseAccessToken;
   }
 
   @Override